Refactor meta test package names
