+++ src/main/java/nl/tudelft/cse1110/andy/execution/metatest/library/evaluators/MetaEvaluator.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cse1110.andy.execution.metatest.evaluators;
+package nl.tudelft.cse1110.andy.execution.metatest.library.evaluators;
 
 public interface MetaEvaluator {
     String evaluate(String oldLibraryCode);
