+++ src/main/java/nl/tudelft/cse1110/andy/execution/metatest/library/evaluators/StringReplacementEvaluator.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cse1110.andy.execution.metatest.evaluators;
+package nl.tudelft.cse1110.andy.execution.metatest.library.evaluators;
 
 public class StringReplacementEvaluator implements MetaEvaluator {
