+++ src/test/java/unit/execution/metatest/LibraryMetaTestTest.java
@@ -1,7 +1,7 @@
 package unit.execution.metatest;
 
 import nl.tudelft.cse1110.andy.config.MetaTest;
-import nl.tudelft.cse1110.andy.execution.metatest.implementations.LibraryMetaTest;
+import nl.tudelft.cse1110.andy.execution.metatest.library.LibraryMetaTest;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
